* Separated callback implementation from FLUID project in the ProjectManager
